+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 - Add ability to review and revoke particular logged in user sessions
 - Add ability to change password from user settings screen
+- Add error logs for background jobs plausible/analytics#4657
 
 ### Removed